# swagger-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_getNotifications",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10, max: 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user notifications (Student or Professor)",
        "tags": ["Notifications"]
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_getUnreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Unread count retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get unread notifications count (Student or Professor)",
        "tags": ["Notifications"]
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markNotificationAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a notification as read (Student or Professor)",
        "tags": ["Notifications"]
      }
    },
    "/api/notifications/mark-all-read": {
      "patch": {
        "operationId": "NotificationsController_markAllNotificationsAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All notifications marked as read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark all notifications as read (Student or Professor)",
        "tags": ["Notifications"]
      }
    },
...
  },
  "info": {
    "title": "AI Professor API",
    "description": "API documentation for AI Professor backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}